+++ src/RuleEngine/Action/Action.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace ILIAS\RuleEngine\Action;
+
+/**
+ * Action
+ *
+ * @author Martin Studer ms@studer-raimann.ch
+ */
+abstract class Action {
+
+	abstract function executeAction();
+
+}
