--- src/RuleEngine/Asserter/Asserter.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace ILIAS\RuleEngine\Asserter;
-
-/**
- * Class Asserter
- *
- * Asserter: evaluate a model representing a rule.
- *
- */
-interface Asserter
-{
-	public function evaluate(): bool;
-}
