+++ src/RuleEngine/Exception/Exception.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace ILIAS\RuleEngine\Exceptions;
+namespace ILIAS\RuleEngine\Exception;
 
 require_once('Services/Exceptions/classes/class.ilException.php');
